+++ packages/sandagent-ui/src/components/artifacts/index.ts
@@ -1,6 +1,0 @@
-export type { ArtifactData, ArtifactItemProps } from "./ArtifactItem";
-export { ArtifactItem } from "./ArtifactItem";
-export type { ArtifactPanelProps } from "./ArtifactPanel";
-export { ArtifactPanel } from "./ArtifactPanel";
-
-export { getFileExtensionFromMimeType } from "./utils";